pyproject.toml
[tool.poetry]
name = "nlp-blueprint"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "nlp_blueprint"}]

[tool.poetry.dependencies]
python = "^3.10"
farm-haystack = {extras = ["faiss"], version = "^1.16.1"}
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
